requirements.txt
Django==3.2.11
djangorestframework==3.13.1
zappa==0.54.1